Adds new composable for event source
